[FIX] split test in two, one for testing without search and another for testing with search
